conversion warns suppressed

+++ src/scag/util/storage/SerialBuffer.cpp
@@ -139,7 +139,7 @@ uint32_t SerialBuffer::ReadInt32()
         throw SerialBufferOutOfBounds();
 
 	uint32_t i;
-	blkread((char*)&i, sizeof(i));
+	blkread((char*)&i, int(sizeof(i)));
     return ntohl(i);
 }
 
@@ -149,7 +149,7 @@ uint16_t SerialBuffer::ReadInt16()
         throw SerialBufferOutOfBounds();
 
 	uint16_t i;
-	blkread((char*)&i, sizeof(i));
+	blkread((char*)&i, int(sizeof(i)));
     return ntohs(i);
 }
 
@@ -159,37 +159,37 @@ uint8_t SerialBuffer::ReadInt8()
         throw SerialBufferOutOfBounds();
 
     uint8_t i;
-	blkread((char*)&i, sizeof(i));
+	blkread((char*)&i, int(sizeof(i)));
     return i;
 }
 
 void SerialBuffer::WriteInt32(uint32_t i)
 {
     i = htonl(i);
-    blkwrite((char*)&i, sizeof(i));
+    blkwrite((char*)&i, int(sizeof(i)));
 }
 
 void SerialBuffer::WriteInt16(uint16_t i)
 {
     i = htons(i);
-    blkwrite((char*)&i, sizeof(i));
+    blkwrite((char*)&i, int(sizeof(i)));
 }
 
 void SerialBuffer::WriteInt8(uint8_t i)
 {
-    blkwrite((char*)&i, sizeof(i));
+    blkwrite((char*)&i, int(sizeof(i)));
 }
 
 void SerialBuffer::WriteString(const char *str)
 {
-    uint16_t len = ::strlen(str);
+    uint16_t len = uint16_t(::strlen(str));
     WriteInt16(len);
     blkwrite(str, len);
 }
 
 void SerialBuffer::WriteString(const wchar_t *str)
 {
-    uint16_t len = std::wcslen(str);
+    uint16_t len = uint16_t(std::wcslen(str));
     WriteInt16(len);
     while(len--)
         WriteInt16((uint16_t)*(str++));
